[Store] Move Indexer and IndexerInterface to Document namespace
Move Indexer.php and IndexerInterface.php from src/store/src/ to src/store/src/Document/ and update their namespaces from Symfony\AI\Store to Symfony\AI\Store\Document. Update all references across the codebase including tests, examples, and bundle configuration.
